+++ Sphere.h
@@ -3,12 +3,12 @@
 
 class Sphere : public Surface {
 public:
-	Sphere(const Vertex center, const float r, Material m);
+	Sphere(const Vertex center, const double r, Material m);
 
 	~Sphere() = default;
 
 	bool intersection(const Direction inDirection, const Vertex start, Vertex& refIntersection) override;
 
 private:
-	float radius;
+	double radius;
 };
